story-engine/backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
google-generativeai==0.8.3
pydantic==2.8.0
python-multipart==0.0.12
python-dotenv==1.0.1
